reinitiating db is not working

+++ database.py
@@ -1,40 +1,56 @@
 #-*-coding:utf-8 -*
 import mysql.connector
-import
+
 import config
 
 class Database:
     def __init__(self):   
-        self.connection = mysql.connector.connect\
+        self.my_database = mysql.connector.connect\
         (host = config.HOST, user = config.USER, password = config.PASSWORD)
-        self.cursor = self.connection.cursor()
         self.status = False
+
+    def open_cursor(self):
+        self.cursor = self.my_database.cursor()
+
+    def close_cursor(self):
+        self.cursor.close()
 
     def check(self):
         querries = ("category","product")
         try: 
             for querry in querries:
+                self.open_cursor()
                 self.cursor.execute("SELECT * FROM p5.%s"% querry)
                 self.cursor.fetchall()
                 if self.cursor.rowcount > 1:
                     self.status = True
+                self.close_cursor()
         except:
             self.status = False
             print ("bug")
 
     def create(self):
         with open(config.SQL_FILE, "r") as file:
+            self.open_cursor()
             content = file.read()
             querries = content.split(";")
             for querry in querries:
                 self.cursor.execute(querry)
+            self.close_cursor()
 
     def delete(self):
-        statement = "DROP DATABASE IF EXISTS p5"
-        self.cursor.execute(statement)
-        self.connection.commit()
+        if self.status == False:
+            self.open_cursor()
+            statement = "DROP DATABASE IF EXISTS p5;"
+            self.cursor.execute(statement)
+            self.my_database.commit()
+            self.close_cursor()
+
+            
+
 
     def insert_categories(self, download):
+        self.open_cursor()
         statement = "INSERT INTO p5.category (id_origin, name,\
         url) VALUES (%s, %s, %s)"
         value = []
@@ -46,9 +62,11 @@ class Database:
             else:
                 pass
         self.cursor.executemany(statement, value)
-        self.connection.commit()
+        self.my_database.commit()
+        self.close_cursor()
 
     def insert_products(self, category, download):
+        self.open_cursor()
         statement = "INSERT INTO p5.product (id_origin, product_name,\
         nutriscore_grade, category_id, url, stores)\
         VALUES (%s, %s, %s, %s, %s, %s)"
@@ -65,5 +83,6 @@ class Database:
                 print(f"The following error occurred: {error}")
                 pass
         self.cursor.executemany(statement, value)
-        self.connection.commit()
+        self.my_database.commit()
+        self.close_cursor()
